refactor: simplify build and packaging of form components
