--- packages/components/form/.gitignore
@@ -1,3 +1,0 @@
-/node_modules
-/pkg
-/coverage
