+++ packages/components/form/bs-form-input.js
@@ -2,28 +2,32 @@
 import { LitElement, html } from 'lit-element';
 import { ifDefined } from 'lit-html/directives/if-defined';
 import { BsFormInputMixin } from './bs-form-input-mixin.js';
-import { BsFormControlSizeCss } from './bs-form-control-size.css.js';
-import { BsFormInputPlaintextCss } from './bs-form-input-plaintext.css.js';
+import { BsFormControlCss } from './css/bs-form-control.css.js';
+import { BsFormControlSizeCss } from './css/bs-form-control-size.css.js';
 import { BsContentRebootCss } from '@lit-element-bootstrap/content';
 
-export class BsFormInputPlaintext extends BsFormInputMixin(LitElement) {
+export class BsFormInput extends BsFormInputMixin(LitElement) {
 
     static get styles() {
         return [
             BsContentRebootCss,
-            BsFormControlSizeCss,
-            BsFormInputPlaintextCss
+            BsFormControlCss,
+            BsFormControlSizeCss
         ];
     }
 
     render() {
         return html`
-            <input 
-                type="text"
+            <input
+                class="form-control"
+                .type=${this.type}
                 .value=${this.value}
                 name=${ifDefined(this.name)}
                 maxlength=${ifDefined(this.maxlength)}
                 minlength=${ifDefined(this.minlength)}
+                max=${ifDefined(this.max)}
+                min=${ifDefined(this.min)}
+                step=${ifDefined(this.step)}
                 pattern=${ifDefined(this.pattern)}
                 placeholder=${ifDefined(this.placeholder)}
                 ?readonly=${this.readonly}
@@ -33,5 +37,5 @@ export class BsFormInputPlaintext extends BsFormInputMixin(LitElement) {
     }
 };
 
-if(!window.customElements.get('bs-form-input-plaintext'))
-    window.customElements.define('bs-form-input-plaintext', BsFormInputPlaintext);
+if(!window.customElements.get('bs-form-input'))
+    window.customElements.define('bs-form-input', BsFormInput);
