+++ packages/components/form/css/bs-form-checkbox-input.css.js
