+++ packages/components/form/index.js
@@ -1,32 +1,20 @@
+export * from './css'
 export { BsForm } from './bs-form.js';
-export { BsFormCss } from './bs-form.css.js';
 export { BsFormSelect } from './bs-form-select.js';
-export { BsFormSelectCss } from './bs-form-select.css.js';
 export { BsFormRow } from './bs-form-row.js';
-export { BsFormRowCss } from './bs-form-row.css.js';
 export { BsFormRange } from './bs-form-range.js';
 export { BsFormRadioGroup } from './bs-form-radio-group.js';
 export { BsFormRadioInput } from './bs-form-radio-input.js';
-export { BsFormRadioInputCss } from './bs-form-radio-input.css.js';
 export { BsFormLabel } from './bs-form-label.js';
-export { BsFormLabelCss } from './bs-form-label.css.js';
 export { BsFormInput } from './bs-form-input.js';
 export { BsFormInputPlaintext } from './bs-form-input-plaintext.js';
-export { BsFormInputPlaintextCss } from './bs-form-input-plaintext.css.js';
 export { BsFormInputMixin } from './bs-form-input-mixin.js';
 export { BsFormInputFeedback } from './bs-form-input-feedback.js';
-export { BsFormInputFeedbackCss } from './bs-form-input-feedback.css.js';
 export { BsFormHelpText } from './bs-form-help-text.js';
 export { BsFormGroup } from './bs-form-group.js';
-export { BsFormGroupCss } from './bs-form-group.css.js';
 export { BsFormFileInput } from './bs-form-file-input.js';
 export { BsFormFieldset } from './bs-form-fieldset.js';
 export { BsFormCheckboxInput } from './bs-form-checkbox-input.js';
-export { BsFormCheckboxInputCss } from './bs-form-checkbox-input.css.js';
 export { BsFormCheckLabel } from './bs-form-check-label.js';
-export { BsFormCheckLabelCss } from './bs-form-check-label.css.js';
 export { BsFormCheckGroup } from './bs-form-check-group.js';
-export { BsFormCheckGroupCss } from './bs-form-check-group.css.js';
 export { BsFormTextarea } from './bs-form-textarea.js';
-export { BsFormControlSizeCss } from './bs-form-control-size.css.js';
-export { BsFormControlCss } from './bs-form-control.css.js';
