+++ packages/components/form/package.json
@@ -2,8 +2,8 @@
   "name": "@lit-element-bootstrap/form",
   "version": "1.0.0-alpha.4",
   "description": "Bootstrap v4.1.3 form components in LitElement",
-  "main": "src/index.js",
-  "module": "src/index.js",
+  "main": "index.js",
+  "module": "index.js",
   "type": "module",
   "author": "Nikolas Christou",
   "homepage": "https://github.com/nik-christou/lit-element-bootstrap",
@@ -12,9 +12,24 @@
     "url": "https://github.com/nik-christou/lit-element-bootstrap"
   },
   "license": "MIT",
+  "files": [
+    "index.js",
+    "index.d.ts",
+    "global.d.ts",
+    "bs-*.js",
+    "bs-*.d.ts",
+    "css/*.js",
+    "css/*.d.ts",
+    "custom-elements.json"
+  ],
   "scripts": {
-    "pika:build": "pika build",
+    "prepack": "yarn build",
+    "build": "npm run wca-analyze:json",
     "serve:demo": "es-dev-server --node-resolve --app-index demo.html --preserve-symlinks",
+    "lit-analyze": "lit-analyzer bs-*.js css/*.js --failFast",
+    "wca-analyze:json": "wca analyze . --outFile custom-elements.json",
+    "lint:prettier": "prettier \"./**/*.js\" --tab-width 4 --list-different",
+    "format:prettier": "prettier \"./**/*.js\" --tab-width 4 --write",
     "test": "karma start",
     "test:watch": "karma start --auto-watch=true --single-run=false",
     "test:update-snapshots": "karma start --update-snapshots",
@@ -35,29 +50,11 @@
     "@open-wc/karma-esm": "^2.9.8",
     "@open-wc/testing": "^2.3.9",
     "@open-wc/testing-karma": "^3.1.54",
-    "@pika/pack": "^0.5.0",
-    "@pika/plugin-build-web": "^0.7.1",
-    "@pika/plugin-bundle-web": "^0.7.1",
-    "@pika/plugin-standard-pkg": "^0.7.1",
     "es-dev-server": "^1.23.1",
     "deepmerge": "^4.2.2",
     "karma": "^4.4.1",
-    "sinon": "^7.5.0"
-  },
-  "@pika/pack": {
-    "pipeline": [
-      [
-        "@pika/plugin-standard-pkg"
-      ],
-      [
-        "@pika/plugin-build-web"
-      ],
-      [
-        "@pika/plugin-bundle-web",
-        {
-            "entrypoint": "unpkg"
-        }
-      ]
-    ]
+    "lit-analyzer": "^1.1.9",
+    "sinon": "^7.5.0",
+    "web-component-analyzer": "^0.1.20"
   }
 }
